migrated files to ts, removed unused functions from lodash_extended
